+++ spec/integration/braintree/transaction_gateway_spec.js
@@ -418,6 +418,7 @@ describe('TransactionGateway', function () {
               unitTaxAmount: '1.23',
               unitOfMeasure: 'gallon',
               discountAmount: '1.02',
+              taxAmount: '1.24',
               totalAmount: '45.15',
               productCode: '23434',
               commodityCode: '9SAASSD8724',
@@ -440,6 +441,7 @@ describe('TransactionGateway', function () {
             assert.equal(lineItem.unitTaxAmount, '1.23');
             assert.equal(lineItem.unitOfMeasure, 'gallon');
             assert.equal(lineItem.discountAmount, '1.02');
+            assert.equal(lineItem.taxAmount, '1.24');
             assert.equal(lineItem.totalAmount, '45.15');
             assert.equal(lineItem.productCode, '23434');
             assert.equal(lineItem.commodityCode, '9SAASSD8724');
@@ -462,6 +464,7 @@ describe('TransactionGateway', function () {
               unitAmount: '45.1232',
               unitOfMeasure: 'gallon',
               discountAmount: '1.02',
+              taxAmount: '0.11',
               totalAmount: '45.15',
               productCode: '23434',
               commodityCode: '9SAASSD8724'
@@ -492,6 +495,7 @@ describe('TransactionGateway', function () {
             assert.equal(lineItem1.unitAmount, '45.1232');
             assert.equal(lineItem1.unitOfMeasure, 'gallon');
             assert.equal(lineItem1.discountAmount, '1.02');
+            assert.equal(lineItem1.taxAmount, '0.11');
             assert.equal(lineItem1.totalAmount, '45.15');
             assert.equal(lineItem1.productCode, '23434');
             assert.equal(lineItem1.commodityCode, '9SAASSD8724');
@@ -507,6 +511,7 @@ describe('TransactionGateway', function () {
             assert.equal(lineItem2.unitOfMeasure, 'gallon');
             assert.equal(lineItem2.totalAmount, '10.10');
             assert.equal(lineItem2.discountAmount, null);
+            assert.equal(lineItem2.taxAmount, null);
             assert.equal(lineItem2.productCode, null);
             assert.equal(lineItem2.commodityCode, null);
             done();
@@ -1460,6 +1465,90 @@ describe('TransactionGateway', function () {
         specHelper.defaultGateway.transaction.sale(transactionParams, function (err, response) {
           assert.isFalse(response.success, 'response had no errors');
           assert.equal(response.errors.for('transaction').for('lineItems').for('index1').on('unitTaxAmount')[0].code, ValidationErrorCodes.Transaction.LineItem.UnitTaxAmountMustBeGreaterThanZero);
+          done();
+        });
+      });
+
+      it('handles validation error tax amount format is invalid', function (done) {
+        let transactionParams = {
+          type: 'sale',
+          amount: '35.05',
+          paymentMethodNonce: Nonces.AbstractTransactable,
+          lineItems: [
+            {
+              quantity: '1.2322',
+              name: 'Name #2',
+              kind: 'credit',
+              unitAmount: '45.0122',
+              unitOfMeasure: 'gallon',
+              discountAmount: '1.02',
+              taxAmount: '2.012',
+              totalAmount: '45.15',
+              productCode: '23434',
+              commodityCode: '9SAASSD8724'
+            }
+          ]
+        };
+
+        specHelper.defaultGateway.transaction.sale(transactionParams, function (err, response) {
+          assert.isFalse(response.success, 'response had no errors');
+          assert.equal(response.errors.for('transaction').for('lineItems').for('index0').on('taxAmount')[0].code, ValidationErrorCodes.Transaction.LineItem.TaxAmountFormatIsInvalid);
+          done();
+        });
+      });
+
+      it('handles validation error tax amount is too large', function (done) {
+        let transactionParams = {
+          type: 'sale',
+          amount: '35.05',
+          paymentMethodNonce: Nonces.AbstractTransactable,
+          lineItems: [
+            {
+              quantity: '1.2322',
+              name: 'Name #2',
+              kind: 'credit',
+              unitAmount: '45.0122',
+              unitOfMeasure: 'gallon',
+              discountAmount: '1.02',
+              taxAmount: '2147483648',
+              totalAmount: '45.15',
+              productCode: '23434',
+              commodityCode: '9SAASSD8724'
+            }
+          ]
+        };
+
+        specHelper.defaultGateway.transaction.sale(transactionParams, function (err, response) {
+          assert.isFalse(response.success, 'response had no errors');
+          assert.equal(response.errors.for('transaction').for('lineItems').for('index0').on('taxAmount')[0].code, ValidationErrorCodes.Transaction.LineItem.TaxAmountIsTooLarge);
+          done();
+        });
+      });
+
+      it('handles validation error tax amount must be greater than zero', function (done) {
+        let transactionParams = {
+          type: 'sale',
+          amount: '35.05',
+          paymentMethodNonce: Nonces.AbstractTransactable,
+          lineItems: [
+            {
+              quantity: '1.2322',
+              name: 'Name #2',
+              kind: 'credit',
+              unitAmount: '45.0122',
+              unitOfMeasure: 'gallon',
+              discountAmount: '1.02',
+              taxAmount: '-1.23',
+              totalAmount: '45.15',
+              productCode: '23434',
+              commodityCode: '9SAASSD8724'
+            }
+          ]
+        };
+
+        specHelper.defaultGateway.transaction.sale(transactionParams, function (err, response) {
+          assert.isFalse(response.success, 'response had no errors');
+          assert.equal(response.errors.for('transaction').for('lineItems').for('index0').on('taxAmount')[0].code, ValidationErrorCodes.Transaction.LineItem.TaxAmountMustBeGreaterThanZero);
           done();
         });
       });