+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 * Add `sourceMerchantId` to `WebhookNotification`s if present
+* Add support for taxAmount field on transaction lineItems
 
 ## 2.5.0
 * Add support for Level 3 summary parameters: `shippingAmount`, `discountAmount`, and `shipsFromPostalCode`